requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
python-multipart
PyPDF2